+++ src/Bridge/PhpTerm/PhpTermBackend.php
@@ -17,15 +17,17 @@ use PhpTui\Term\Colors;
 use PhpTui\Term\Event\CursorPositionEvent;
 use PhpTui\Term\Size;
 use PhpTui\Term\Terminal as PhpTermTerminal;
-use PhpTui\Tui\Model\AnsiColor;
 use PhpTui\Tui\Model\Area;
 use PhpTui\Tui\Model\Backend;
 use PhpTui\Tui\Model\BufferUpdates;
 use PhpTui\Tui\Model\ClearType;
 use PhpTui\Tui\Model\Color;
+use PhpTui\Tui\Model\Color\AnsiColor;
+use PhpTui\Tui\Model\Color\LinearGradient;
+use PhpTui\Tui\Model\Color\RgbColor;
 use PhpTui\Tui\Model\Modifier;
 use PhpTui\Tui\Model\Position;
-use PhpTui\Tui\Model\RgbColor;
+use PhpTui\Tui\Model\Widget\FractionalPosition;
 use RuntimeException;
 
 class PhpTermBackend implements Backend
@@ -226,6 +228,11 @@ class PhpTermBackend implements Backend
         }
         if ($color instanceof RgbColor) {
             return new SetRgbForegroundColor($color->r, $color->g, $color->b);
+        }
+
+        // if we have a raw gradient, use it's first stop.
+        if ($color instanceof LinearGradient) {
+            return $this->setForegroundColor($color->at(FractionalPosition::at(0, 0)));
         }
 
         throw new RuntimeException(sprintf('Do not know how to set color of type "%s"', $color::class));